requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
discord.py==2.5.2
Flask==3.1.1
frozenlist==1.6.2
git-filter-repo==2.47.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
propcache==0.3.1
Werkzeug==3.1.3
yarl==1.20.0
